src/requirements.txt
requests==2.31.0
markdownify==0.12.6
python-dotenv==1.0.0